style: Modal style 수정
